[UI] Extracted webpack configuration into separate files
